backend/requirements.txt
# =========================
# FASTAPI BACKEND
# =========================
fastapi==0.115.0
uvicorn==0.30.6

# =========================
# DATABASE
# =========================
pymongo==4.8.0
python-dotenv==1.0.1
dnspython==2.6.1

# =========================
# AUTHENTICATION
# =========================
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.1.1

# =========================
# AI / ML
# =========================
numpy==1.26.4
scikit-learn==1.5.1

# =========================
# HTTP / NETWORKING
# =========================
httpx==0.27.0
requests==2.32.3

# =========================
# UTILITIES
# =========================
pydantic==2.9.2
pydantic[email]==2.9.2

# =========================
# NOTIFICATIONS
# =========================
twilio==9.2.3
